% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/love_cats.R
\name{love_cats}
\alias{love_cats}
\title{love_cats}
\usage{
love_cats(x)
}
\arguments{
\item{x}{a numeric value as input}
}
\value{
character string or statement
}
\description{
Are you a cat love or dog lover?
}
\examples{
love_cats(x=10)
}
